# public/swagger-api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Query Report Builder API",
    "description": "API untuk membuat, mengelola, dan mengeksekusi query report secara dinamis. API ini memungkinkan user untuk membuat report dengan parameter, menyimpannya, dan mengeksekusi kapan saja dengan parameter yang berbeda.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@queryreport.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development Server (Local)"
    },
    {
      "url": "https://192.168.30.70:8080",
      "description": "Development Server (Network)"
    }
  ],
...
  "paths": {
    "/api/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login pengguna",
        "description": "Endpoint untuk autentikasi user dengan username dan password. Session akan dibuat setelah login berhasil.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "admin",
                    "description": "Username pengguna"
                  },
                  "password": {
                    "type": "string",
                    "example": "admin123",
                    "description": "Password pengguna"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "object",
                      "properties": {
                        "ID": {
                          "type": "string",
                          "example": "PTG001"
                        },
                        "UID": {
                          "type": "string",
                          "example": "16156151515"
                        },
                        "Nama": {
                          "type": "string",
                          "example": "Muhammad Nur Amin"
                        },
                        "Organisasi": {
                          "type": "integer",
                          "example": 787878
                        },
                        "HakAkses": {
                          "type": "object",
                          "properties": {
                            "reporting": {
                              "type": "array",
                              "items": {
                                "type": "object"
                              }
                            }
                          }
                        }
                      }
                    },
                    "metadata": {
                      "$ref": "#/components/schemas/Metadata"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Username atau password tidak diisi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Username atau password salah",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profile": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get profil pengguna yang sedang login",
        "description": "Endpoint untuk mendapatkan informasi profil pengguna yang sedang login. Memerlukan session aktif.",
        "responses": {
          "200": {
            "description": "Berhasil mendapatkan profil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "$ref": "#/components/schemas/UserProfile"
                    },
                    "metadata": {
                      "$ref": "#/components/schemas/Metadata"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - belum login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Metadata": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK",
            "description": "Status message"
          },
          "code": {
            "type": "integer",
            "example": 200,
            "description": "HTTP status code"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "response": {
            "nullable": true,
            "example": null,
            "description": "Response data (null untuk error)"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Error message",
                "description": "Pesan error"
              },
              "code": {
                "type": "integer",
                "example": 400,
                "description": "HTTP status code"
              }
            }
          }
        }
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "ID": {
            "type": "string",
            "example": "PTG001",
            "description": "User ID"
          },
          "UID": {
            "type": "string",
            "example": "16156151515",
            "description": "Unique ID"
          },
          "Nama": {
            "type": "string",
            "example": "Muhammad Nur Amin",
            "description": "Nama lengkap user"
          },
          "Organisasi": {
            "type": "integer",
            "example": 787878,
            "description": "ID Organisasi"
          },
          "HakAkses": {
            "type": "object",
            "description": "Hak akses user",
            "properties": {
              "reporting": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
    }
  }
}